Module Par_interp.Annotate

exception Empty_par_prg_block

Raised whenever a Par statement contains an empty program.

exception Empty_prg_block

Raised if the program to be annotated is empty.

val ann_prg_in_par : Ast.program -> Ast_ann.program_ann

Annotates a program inside a Par block. This is an auxiliary function for ann_stmt, so refer to its description for details.

val ann_stmt : s:Ast.stmt -> Ast_ann.stmt_ann

Annotates a statement.

An int list (which will keep track of statements' execution order) is added to each statement.

As for the Par statement:

In addition to the previously mentioned int list added to all statements, an int scalar to keep track of the number of child threads which finished the execution of the programs inside the Par statement is added as well;

The statements contained in its two program blocks will be recursively annotated as well, and Par_prg_start - Par_prg_end statements are added at the beginning and at the end of the program respectively, in order to define the programs' boundaries.

parameter s

The statement to be annotated.

returns

The annotated statement.

raises Empty_par_prg_block

in case a program contained inside a Par block is empty.

val ann_prg : prg:Ast.program -> Ast_ann.program_ann

Annotates a program (by representing it as a zipper structure, among the other things) and its statements.

Also, puts Program_start statement at the beginning and Program_end at the end of the program, in order to set the whole program's boundaries.

parameter prg

The program to be annotated.

returns

The annotated program.

raises Empty_prg_block

when the program to annotate (prg) is empty.