Module Par_interp.Delta

exception Delta_stack_error of string

val exc_ide_stack_not_found : string -> exn

type ide_stacks =

| Ide_stacks of (Ast.ide * int list) list

ide_stacks is a type representing a list of tuples (var_name, prev_values_stack); its goal is to keep track of variables' previous values each time a destructive assignment is performed, by pushing the older value onto the respective variable's stack before performing the assignment. By doing so, previous values of any given variable can be retrieved during reverse execution by simply "popping" the value in the head of the list and assigning it to that variable in the sigma store.

val private_push_ide : ide_stk:ide_stacks -> ide_name:Ast.ide -> value:int -> ide_stacks

Pushes a value onto the specified ide_name variable's stack.

If a tuple (ide_name, ide_stack_list) already exists in ide_stk, value is simply prepended to that list (pushed onto that ide_stack_list): (ide_name, value :: ide_stack_list).

If a tuple for ide_name doesn't exist, a tuple (ide_name, value :: ) gets created and appended to ide_stk.

parameter ide_stk

The list of tuples (variable_name, variable_name_stack_list) representing variables' auxiliary store.

parameter ide_name

The name of the variable which underwent a destructive assignment, for which we want to preserve its previous value by pushing it onto its associated stack.

parameter value

The previous integer value to push onto the stack associated to ide_name.

returns

The updated ide_stk, where value has been pushed onto ide_name's associated stack.

val private_top_ide : ide_stk:ide_stacks -> ide_name:Ast.ide -> int

Retrieves the integer value from the top of the specified ide_name variable's stack.

parameter ide_stk

The list of tuples (variable_name, variable_name_stack_list).

parameter ide_name

The name of the variable associated to the stack from which we want to retrieve the integer value on its top.

returns

The integer on the top of the stack associated to ide_name.

raises Delta_stack_error

with a descriptive message about the variable/identifier, in case the tuple (ide_name, ide_name_stack_list) wasn't found inside ide_stk.

val private_pop_ide : ide_stk:ide_stacks -> ide_name:Ast.ide -> ide_stacks

Pops (removes) the integer value from the top of the specified ide_name variable's stack.

parameter ide_stk

The list of tuples (variable_name, variable_name_stack_list).

parameter ide_name

The name of the variable associated to the stack from which we want to retrieve the integer value on its top.

returns

The updated ide_stk, where the element previously located on top of ide_name's associated stack has been removed.

If ide_name's associated stack contained a single element, the whole (ide_name, ide_name_stack_list) pair gets removed from ide_stk.

raises Delta_stack_error

with a descriptive message about the variable/identifier, in case the tuple (ide_name, ide_name_stack_list) wasn't found inside ide_stk.

type delta =

| Delta of ide_stacks

delta is a type representing the auxiliary store, consisting of:

A list of stacks ide_stacks, where each stack is contained inside a pair (<identifier>, <integer list>) that keeps track of each variable's value: every time a destructive assignment is performed on a certain <ide> variable, its previous value is pushed onto the stack associated to that <ide> variable, in order to preserve it.

Unlike the serial version, if_stack and while_stack aren't present, since the parallel version doesn't implement neither Ifthenelse conditional execution nor While iterations (yet).

val push_ide : d:delta -> ide_name:Ast.ide -> value:int -> delta

This is a wrapper which applies private_push_ide to the parameter ide_stks inside the specified d delta parameter.

val top_ide : d:delta -> ide_name:Ast.ide -> int

This is a wrapper which applies private_top_ide to the parameter ide_stks inside the specified d delta parameter.

val pop_ide : d:delta -> ide_name:Ast.ide -> delta

This is a wrapper which applies private_pop_ide to the parameter ide_stks inside the specified d delta parameter.

val empty_delta : delta

Returns an empty auxiliary store,