Module Par_interp.Interpreter

exception Assignment_to_non_var of string

exception Illegal_statement_fwd_execution of string

exception Illegal_statement_rev_execution of string

val sem_int : expr:Ast.int_expr -> s:Sigma.sigma -> int

Integer semantics' definition.

parameter expr

The integer expression to evaluate.

parameter s

The sigma store (used in case the expression contains variables whose value must be retrieved.)

returns

The integer value resulting from the expression's evaluation.

val assign_var_fwd : e1:Ast.int_expr -> e2:Ast.int_expr -> state:State.state -> State.state

Performs forward execution's destructive assignment:

Evaluates e1, making sure it's a variable;

Evaluates e2;

Assigns e2's evaluation to the variable specified by e1, and returns the state with the above changes applied (see State.dassign_fwd for details such as pushing the previous value in the delta store before performing the assignment.)

parameter e1

The expression containing the variable to which we must assign the new value.

parameter e2

The expression containing the value to assign.

parameter state

The state where the assignment must be performed.

returns

The new state where the assignment has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val assign_var_rev : e1:Ast.int_expr -> state:State.state -> State.state

Performs reverse execution's destructive assignment:

Evaluates e1, making sure it's a variable ide;

Assigns ide's previous value to ide and returns the state with the above changes applied (see State.dassign_rev for details such as popping ide's previous value from the delta store.)

parameter e1

The expression containing the variable to which we must assign the previous value.

parameter state

The state where the assignment must be performed.

returns

The new state where the assignment has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val cadd : e1:Ast.int_expr -> e2:Ast.int_expr -> state:State.state -> State.state

Performs constructive sum:

Evaluates e1, making sure it's a variable;

Evaluates e2;

Sums e2's evaluation to the variable specified by e1, and returns the state with the above changes applied (see State.cadd for details.)

parameter e1

The expression containing the variable to which we must add e2's evaluation

parameter e2

The expression containing the value to add to the variable specified by e1.

parameter state

The state where the sum must be performed.

returns

The new state where the sum has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val csub : e1:Ast.int_expr -> e2:Ast.int_expr -> state:State.state -> State.state

Performs constructive subtraction:

Evaluates e1, making sure it's a variable;

Evaluates e2;

Subtracts e2's evaluation from the variable specified by e1, and returns the state with the above changes applied (see State.csub for details.)

parameter e1

The expression containing the variable from which we must subtract e2's evaluation

parameter e2

The expression containing the value to subtract from the variable specified by e1.

parameter state

The state where the subtraction must be performed.

returns

The new state where the subtraction has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val exec_par_fwd : Ast_ann.program_ann -> Ast_ann.program_ann -> int -> State.state -> State.state

Evaluates the Par statement in forward execution mode.

After annotating the Par statement with the current statement counter, the thread where the Par statement has been encountered is moved to the list of waiting threads and two new child threads (each containing a program taken from the Par statement) are created in the list of running threads: the resulting state is returned.

val exec_par_rev : Ast_ann.program_ann -> int -> State.state -> State.state

Evaluates the Par statement in reverse execution mode.

Basically, what's happening here is that the thread where the Par statement has been encountered is moved back to the waiting threads, and the child thread which terminated last is re-created by retrieving its program from the Par statement itself.

val sem_stmt_fwd : int -> State.state -> State.state

Given a thread ID and a state, performs a single evaluation step in forward execution mode in the program associated to the given thread ID, and returns the resulting state.

val sem_stmt_rev : State.state -> State.state

Given a state, performs a single evaluation step in reverse execution mode by retrieving the last statement previously executed in forward execution mode, and returns the resulting state.

val sem_prg_fwd : int -> State.state -> State.state

Given a thread ID and a state, evaluates all statements in forward execution mode until the end of the program contained in the specified thread, and returns the resulting state (in other words, it executes the whole program contained in the specified thread.)

val sem_prg_rev : State.state -> State.state

Given a state, evaluates all statements until the beginning of the program in reverse execution mode and returns the resulting state.

val sem_prg_fwd_steps : int -> int -> State.state -> State.state

Given a thread ID tid, a specified num_steps integer and a state curr_state, let remaining_stmts be the number of statements between the current statement in tid's program and the Program_end / Par_prg_end boundary statement; then the function performs min(remaining_stmts, num_steps) statement evaluations in forward execution mode and returns the resulting state.

If num_steps <= 0, curr_state is returned unaltered.

val sem_prg_rev_steps : int -> State.state -> State.state

Given a specified num_steps integer and a state curr_state, let remaining_stmts be the number of statements between the current statement in curr_state's program and the Program_start boundary statement; then the function performs min(remaining_stmts, num_steps) statement evaluations in reverse execution mode and returns the resulting state.

If num_steps <= 0, curr_state is returned unaltered.