Module Par_interp.State

exception Last_executed_stmt_not_found

exception Last_executed_thread_not_found

type state =

| State of Thread.thread list * Thread.thread list * int * int * Sigma.sigma * Delta.delta

state is a type consisting of a tuple including:

running_threads (the list of threads currently executing);

waiting_threads (the list of threads that generated the threads in running_threads, waiting for their children to finish their execution before resuming their own execution by moving back to running_threads;

num_curr_stmt: an integer counting the number of statements executed so far, used to annotate statements as they are executed;

num_curr_thread: an integer counting the number of created threads (both active and killed), used to assign an ID to each thread upon creation; whenever the root thread reaches its program's last statement (or the first one, if in reverse execution), this value is reset to the initial value (root_tid_value + 1, that is.)

The program store (sigma);

The auxiliary store (delta).

val first_stmt_value : int

Value used to initialize state's num_curr_stmt value.

val root_tid_value : int

Value used to set the root thread ID, and to initialize state's num_curr_thread value (root_tid_value + 1).

val dassign_fwd : ide_name:Ast.ide -> new_value:int -> state -> state

Executes a destructive assignment in forward execution mode, assigning new_value to the variable ide_name.

In order to preserve ide_name's current value, curr_value is retrieved from the sigma store and pushed onto ide_name's stack in the auxiliary delta store; at that point, ide_name's value in the sigma store can be overwritten with new_value, and the resulting new state is returned.

val dassign_rev : ide_name:Ast.ide -> state -> state

Executes a destructive assignment in reverse execution mode.

ide_name's previous value (retrieved_value) is retrieved from ide_name's stack in the auxiliary delta store, and retrieved_value gets popped (removed) from ide_name's stack; at that point, retrieved_value is assigned to ide_name in the sigma store, and the resulting new state is returned.

val cadd : Ast.ide -> int -> state -> state

This is a wrapper which applies Sigma.cadd to the sigma store contained inside the specified state parameter, and returns the resulting state.

val csub : Ast.ide -> int -> state -> state

This is a wrapper which applies Sigma.csub to the sigma store contained inside the specified state parameter, and returns the resulting state.

val get_running_thread : int -> state -> Thread.thread

Returns the running thread with the given ID contained inside the specified state parameter.

val add_running_thread : Thread.thread -> state -> state

Adds a thread to the list of running threads inside the specified state parameter.

val update_running_thread : Thread.thread -> state -> state

Updates (replaces) a thread in the list of running threads inside the specified state parameter.

val new_running_thread : Ast_ann.program_ann -> int -> Program.child_branch -> state -> state

Given a program, a parent thread ID, a branch and a state, creates a new thread containing the given program in the list of running threads inside the specified state parameter.

val remove_running_thread : int -> state -> state

Removes the thread with the given ID from the list of running threads inside the specified state parameter.

val get_waiting_thread : int -> state -> Thread.thread

Returns the waiting thread with the given ID contained inside the specified state parameter.

val add_waiting_thread : Thread.thread -> state -> state

Adds a thread to the list of waiting threads inside the specified state parameter.

val update_waiting_thread : Thread.thread -> state -> state

Updates (replaces) a thread in the list of waiting threads inside the specified state parameter.

val remove_waiting_thread : int -> state -> state

Removes the thread with the given ID from the list of waiting threads inside the specified state parameter.

val set_thread_as_waiting : int -> state -> state

Moves the thread with the given ID from the list of running threads to the list of waiting threads.

val set_thread_as_running : int -> state -> state

Moves the thread with the given ID from the list of waiting threads to the list of running threads.

val get_stmt_counter : state -> int

Returns the statement counter (the number of statements executed thus far) contained in the state passed as a parameter.

val push_stmt_counter : int -> state -> state

Pushes the statement counter in the source code stack of the current instruction in the program contained in the specified running thread ID.

val pop_prev_stmt_counter : int -> state -> state

Pops (removes) the statement counter from the top of the source code stack of the previous instruction in the program contained in the specified running thread ID.

val get_sigma : state -> Sigma.sigma

Returns the sigma store contained inside the specified state parameter.

val init : Ast_ann.program_ann -> Sigma.sigma -> state

Takes a program and an initial sigma store as parameters, and returns a state with

A list of active threads initialized with a root thread containing the given program;

An empty list of waiting threads;

num_curr_stmt initialized to first_stmt_value;

num_curr_thread initialized to root_tid_value + 1;

The given sigma store;

An empty auxiliary (delta) store.

val prev_stmt : int -> state -> state

This is a wrapper which applies Thread.prev_stmt to the thread having the specified running thread ID contained inside the specified state parameter, and returns the resulting state.

val get_prev_stmt : int -> state -> Ast_ann.stmt_ann

This is a wrapper which applies Thread.get_prev_stmt to the thread having the specified running thread ID contained inside the specified state parameter, and returns the resulting state.

val get_curr_stmt : int -> state -> Ast_ann.stmt_ann

This is a wrapper which applies Thread.get_curr_stmt to the running thread with the given ID contained inside the specified state parameter.

val next_stmt : int -> state -> state

This is a wrapper which applies Thread.next_stmt to the running thread with the given ID contained inside the specified state parameter.

val get_child_threads_from_list : int -> state -> Thread.thread * Thread.thread

This is a wrapper which applies Thread.get_child_threads_from_list to the running threads' list contained inside the specified state parameter.

val remove_child_threads_from_list : int -> state -> state

This is a wrapper which applies Thread.remove_child_threads_from_list to the running threads' list contained inside the specified state parameter.

val get_last_executed_thread : state -> Thread.thread

Given a state, searches the thread which performed the last statement execution among the running threads and returns it.

raises Last_executed_thread_not_found

if no thread among the running threads performed the last execution.

val inc_prev_par_finished_children : int -> state -> state

Given a waiting thread ID whose program's previous statement is Par and a state, increments Par's num_chld_done field by 1.

val dec_prev_par_finished_children : int -> state -> state

Given a waiting thread ID whose program's previous statement is Par and a state, decrements Par's num_chld_done field by 1.

val adjust_state_for_rev_semantics : state -> state

Given a state, returns an adjusted state to use before performing a single step in reverse execution mode, so that get_last_executed_thread is guaranteed to find the thread which executed last among the running threads:

If the state's current statement is the first one in the root thread, or the last executed statement is in a thread which is running already, then no adjustment is necessary (the state is returned as-is);

If the last executed statement is in the waiting threads, it means that the last executed instruction was Par: the child threads which have just been created as a result of the Par statement execution must be removed after putting their programs back in Par (in order to guarantee that Par's programs are "rewound" back to the beginning), and the waiting thread must be moved back to the running threads' list;

If the last executed statement is found inside one of the programs contained inside a waiting thread's Par statement, it means a child thread has terminated in the previous statement execution: another thread is created in the running threads, containing the program found inside the waiting thread's Par statement.

raises Last_executed_stmt_not_found

if the last executed statement couldn't be found anywhere.

val inc_num_stmts : state -> state

Given a state, increments num_curr_stmt.

val dec_num_stmts : state -> state

Given a state, decrements num_curr_stmt.

val inc_num_threads : state -> state

Given a state, increments num_curr_thread.

val dec_num_threads : state -> state

Given a state, decrements num_curr_thread.

val reset_num_threads : state -> state

Given a state, resets the num_threads value to the initial value. Used when the program execution reaches the end in forward execution mode or the beginning in reverse execution mode, since at those points we know the only active thread is the root thread (and therefore num_threads can be reset to the value coming after root_tid_value, since it won't overlap with any existing threads' IDs.)

val move_to_next_stmt : int -> state -> state

Given a thread ID and a state, performs the following operations:

Pushes the statement counter onto the current statement's source code stack;

Increments the state's statement counter by 1;

Moves the current instruction of the program contained in the given thread ID to the next instruction;

Returns the updated state.

val move_to_prev_stmt : int -> state -> state

Given a thread ID and a state, performs the following operations:

Pops (removes) the statement counter from the current statement's source code stack;

Decrements the state's statement counter by 1;

Moves the current instruction of the program contained in the given thread ID to the previous instruction;

Returns the updated state.

val handle_finished_par_thread_fwd : tid:int -> state:state -> state

Handles the forward execution of statement Program_ann.Par_prg_end (that is, a child thread terminated its program execution) at state level.

The following steps are performed:

The child thread and the parent thread are retrieved from the state;

The child thread's program is put into the Par statement contained in the parent thread's program, by calling Thread.update_par_prg_fwd;

The child thread is removed from the running threads (since it terminated its execution);

The parent thread is removed from the waiting threads (since it must be replaced by the updated version);

The updated parent thread is added to the state; if all of its child threads finished their execution, the parent thread is placed in the running threads' list and its current program statement advances to the next statement, otherwise it's placed in the waiting threads' list.

parameter tid

The thread ID of the child thread which terminated its execution.

parameter state

The state to update.

returns

The updated state.

val is_prg_at_start : state -> bool

Returns a boolean value indicating whether the program's current statement/instruction in the given state is the main program's first statement/instruction.

val is_thread_at_end : int -> state -> bool

Given a thread ID and a state, returns a boolean value indicating whether the program contained in the given thread is the main program's last statement/instruction.

val is_thread_running : int -> state -> bool

Given a thread ID and a state, returns true if a thread with the given ID is found in the running threads' list, false otherwise.