Module Par_interp.Thread

exception Thread_not_found of string

The string argument indicates the function where the exception has occurred.

type thread =

| Thread of Ast_ann.program_ann * int * int * Program.child_branch

A thread is a tuple containing the following fields:

prg: The annotated program associated to the thread;

tid: An integer value that uniquely identifies the thread;

ptid: An integer value that identifies the parent thread's ID (that is, the tid of the thread that generated this thread);

branch: A variable of type child_branch, indicating whether the thread is a Left child, a Right child, or the Root thread;

val create : Ast_ann.program_ann -> int -> int -> Program.child_branch -> thread

Creates and returns a new thread, initializing it with the given parameters (annotated program, thread id, id of the thread who requested the creation, and branch).

val get_program : thread -> Ast_ann.program_ann

Given a thread, returns the annotated program contained in it.

val get_tid : thread -> int

Given a thread, returns its thread ID.

val get_ptid : thread -> int

Given a thread, returns its parent thread ID.

val get_child_branch : thread -> Program.child_branch

Given a thread, returns its child branch (Left, Right, or Root.)

val update_par_prg : thread -> thread -> thread

Given a child thread and its parent thread, puts the child thread's program into the Par statement contained in the parent thread's program.

The following steps are performed:

The child thread branch is retrieved, in order to determine whether the program contained in the child thread belongs to the first or the second program inside the Program.Par statement of the parent thread;

The Program.Par program branch contained in the parent thread is updated, according to the branch retrieved from the previous step.

parameter t_chld

The child thread containing the program to put inside t_parent's Par statement.

parameter t_parent

The parent thread whose current statement is the Par statement to be updated with the finished child thread's program.

returns

The updated parent thread.

val update_par_prg_fwd : thread -> thread -> thread

Same as update_par_prg, but additionally increments Par's num_chld_done field in the parent program by 1.

val is_children_execution_done : thread -> bool

Given a waiting thread, returns a boolean value indicating whether its children are done with their execution (true) or not (false).

This is a wrapper which applies Program.is_children_execution_done to the program contained inside the specified thread parameter.

val prev_stmt : thread -> thread

This is a wrapper which applies Program.prev_stmt to the program contained inside the specified thread parameter.

val get_prev_stmt : thread -> Ast_ann.stmt_ann

This is a wrapper which applies Program.get_prev_stmt to the program contained inside the specified thread parameter.

val get_curr_stmt : thread -> Ast_ann.stmt_ann

This is a wrapper which applies Program.get_curr_stmt to the program contained inside the specified thread parameter.

val next_stmt : thread -> thread

This is a wrapper which applies Program.next_stmt to the program contained inside the specified thread parameter.

val push_stmt_counter : int -> thread -> thread

This is a wrapper which applies Program.push_stmt_counter to the program contained inside the specified thread parameter.

val top_prev_stmt_counter : thread -> int

This is a wrapper which applies Program.top_stmt_counter to the program contained inside the specified thread parameter.

val pop_prev_stmt_counter : thread -> thread

This is a wrapper which applies Program.pop_prev_stmt_counter to the program contained inside the specified thread parameter.

val get_thread_from_list : int -> thread list -> thread

Given a thread ID and a list of threads, returns the thread in the list matching the ID passed as a parameter, or raises Thread_not_found if no thread ID matches the given one.

val remove_thread_from_list : int -> thread list -> thread list

Given a thread ID and a list of threads, removes the thread from the list matching the ID passed as a parameter, or raises Thread_not_found if no thread ID matches the given one.

val update_thread_in_list : thread -> thread list -> thread list

Given a thread thread and a thread list, updates (replaces) the thread in the list having the same ID as thread with thread.

val get_last_executed_thread_from_list : int -> thread list -> thread option

Given an int representing the number of the last executed statement and a list of threads, returns the thread contained in the given list which was executed last.

val get_last_executed_par_prg_from_list : int -> thread list -> (Ast_ann.program_ann * Program.child_branch * int) option

Given an int value representing the last statement's execution counter and a list of waiting threads, scans the thread list and returns a pair Some (prg, branch, ptid) if the Par statement in one of the threads contains the last executed instruction in one of its two programs (indicated by branch), None otherwise.

val get_child_threads_from_list : int -> thread list -> thread * thread

Given an int representing the parent thread ID and a list of threads, returns the threads generated by the given parent thread ID from the list as a pair (thread1, thread2).

val remove_child_threads_from_list : int -> thread list -> thread list

Given an int representing the parent thread ID and a list of threads, removes the threads generated by the given parent thread ID from the list and returns it.

val inc_prev_par_finished_children : thread -> thread

This is a wrapper which applies Program.inc_prev_par_finished_children to the program contained inside the specified thread parameter.

val dec_prev_par_finished_children : thread -> thread

This is a wrapper which applies Program.dec_prev_par_finished_children to the program contained inside the specified thread parameter.

val is_thread_in_list : int -> thread list -> bool

Given a thread ID and a list of threads, returns true if a thread having as ID the same value specified in the parameter is found inside the list, false otherwise.