Module Par_interp__Program

Implementation of the functions acting on Ast_ann.program_ann.

exception No_previous_statements

exception No_next_statements

exception Statement_without_stack of string

exception Statement_empty_stack of string

exception Root_thread_par_update

exception Prev_stmt_not_Par

type child_branch =

| Left

| Right

| Root

type used to discern whether the thread contains a program contained in the first (Left) or in the second (Right) element of the Par statement of the parent thread that created it.

val max_num_chld_threads : int

Each thread can have max 2 running child threads at a time, since a Par statement creates two child threads and the parent thread waits for both children to complete before proceeding its execution.

val prepend_stmt : stmt_ann:Par_interp.Ast_ann.stmt_ann -> Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program, adds stmt_ann as the program's first statement/instruction.

val append_stmt : stmt_ann:Par_interp.Ast_ann.stmt_ann -> Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program, adds stmt_ann as the program's last statement/instruction.

val prev_stmt : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program prg, returns:

A new program where prg's previous statement becomes the current statement, or

prg, if prg's current statement is the first one (there are no previous statements.)

val next_stmt : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program prg, returns:

A new program where prg's next statement becomes the current statement, or

prg, if prg's current statement is the last one (there are no next statements.)

val last_stmt : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program prg, returns a new program where prg's last statement becomes the current statement.

val get_prev_stmt : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.stmt_ann

Returns the statement preceding the current statement inside the program.

raises No_previous_statements

if the previous statements' list is empty (e.g. the given program's current statement is the first one.)

val get_curr_stmt : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.stmt_ann

Returns the current statement inside the program.

val push_stmt_counter : int -> Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given an int value stmt_counter representing the current statement counter and a program prg, pushes stmt_counter onto the source code stack associated to prg's current statement.

raises Statement_without_stack

if the current statement doesn't have an associated source code stack.

val top_prev_stmt_counter : Par_interp.Ast_ann.program_ann -> int

Given a program prg, returns the statement counter value located on the top of the source code stack associated to prg's previous statement.

raises Statement_without_stack

if the current statement doesn't have an associated source code stack.

raises Statement_empty_stack

if the previous statement's source code stack is empty.

raises No_previous_statements

if there are no statements preceding the current statement.

val pop_prev_stmt_counter : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program prg, pops (removes) the statement counter value from the top of the source code stack associated to prg's previous statement.

raises Statement_without_stack

if the current statement doesn't have an associated source code stack.

raises Statement_empty_stack

if the previous statement's source code stack is empty.

raises No_previous_statements

if there are no statements preceding the current statement.

val is_par_prg_at_end : Par_interp.Ast_ann.program_ann -> bool

Given a program, returns true if the current statement is Par_prg_end, false otherwise.

val inc_prev_par_finished_children : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program having Par as the previous statement, increases num_chld_done field contained in the aforementioned Par statement by 1. Called whenever a new child thread associated to said Par statement is created.

val dec_prev_par_finished_children : Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a program having Par as the previous statement, increases num_chld_done field contained in the aforementioned Par statement by 1. Called whenever a child thread associated to said Par statement terminates.

raises Prev_stmt_not_Par

if the previous statement in the given program parameter is not Par.

val update_par_prg : chld_branch:child_branch -> chld_prg:Par_interp.Ast_ann.program_ann -> parent_prg:Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Given a thread branch, a child program and a parent program whose previous statement is a Par statement, puts the updated child program into the parent's Par statement.

parameter chld_branch

The child thread's branch (Left or Right), to determine whether the program contained in the child thread belongs to the first or the second program inside the Program_ann.Par statement of the parent thread. If it's Root, then an exception is raised.

parameter chld_prg

The updated child program to put into the parent program's Par statement.

parameter t_parent

The parent program whose previous statement is the Par statement to update with the finished child thread's program.

returns

The updated parent program.

raises Root_thread_par_update

if chld_branch is Root.

raises Prev_stmt_not_Par

if the parent program's previous statement is not Par.

val update_par_prg_fwd : chld_branch:child_branch -> chld_prg:Par_interp.Ast_ann.program_ann -> parent_prg:Par_interp.Ast_ann.program_ann -> Par_interp.Ast_ann.program_ann

Handles the forward execution of statement Program_ann.Par_prg_end (that is, a child thread terminated its program execution) by calling update_par_prg and incrementing Par's field num_chld_done by 1.

val is_children_execution_done : Par_interp.Ast_ann.program_ann -> bool

Given a waiting thread's program, returns a boolean value indicating whether its children are done with their execution (true) or not (false).

raises Prev_stmt_not_Par

if the parent program's previous statement is not Par (and therefore we're most likely not dealing with a waiting thread).

val is_prev_stmt_par : Par_interp.Ast_ann.program_ann -> bool

Given a waiting thread's program, returns a boolean value indicating whether its previous statement is Par (true) or not (false).

val get_last_executed_par_prg : int -> Par_interp.Ast_ann.program_ann -> (Par_interp.Ast_ann.program_ann * child_branch) option

Given an int value representing the last statement's execution counter and a program whose previous statement is Par, returns a pair Some (prg, branch) if the Par statement contains the last executed statement in one of its two programs (indicated by branch), None otherwise.

raises Prev_stmt_not_Par

if the given program's previous statement is not Par.