Module Par_interp__Sigma

Implementation of the program store and the functions acting on it.

exception Unbound_variable of string

val exc_unbound_var : string -> exn

type sigma =

| Sigma_empty

| Sigma of Par_interp.Ast.ide * int * sigma

sigma is a type representing the store (that is, the set of all declared identifiers(variables) and the respective integer values associated to them).

val dassign : s:sigma -> ide_name:Par_interp.Ast.ide -> value:int -> sigma

Destructive assignment.

By destructive assignment we mean an assignment where there's no way to retrieve the previous value without additional information; for example, after executing x = 5 we have no way to determine which value was stored inside x without having saved it somewhere before the assignment (that "somewhere" is the delta store.)

Given a sigma store, overwrites the older value associated to ide_name with the new value specified.

parameter s

The sigma store where the variable to be updated is stored.

parameter ide_name

The variable name to update in the sigma store.

parameter value

The new integer value to assign to the ide_name variable.

returns

The updated sigma store, where the identifier ide_name is now associated to value.

raises Unbound_variable

with a descriptive message about the variable/identifier, in case the specified ide_name wasn't found inside s.

val cassign : s:sigma -> ide_name:Par_interp.Ast.ide -> operation:(int -> int) -> sigma

Constructive assignment.

By constructive assignment we mean an assignment where the previous value can be retrieved by looking at the instruction, without any additional information; for example, after executing x += 2 we can retrieve x's value prior to the assignment by trivially subtracting 2 from x.

Given a sigma store, updates the older value associated to ide_name in the way specified by the operation parameter function.

parameter s

The sigma store where the variable to be updated is stored.

parameter ide_name

The variable name to update in the sigma store.

parameter operation

A function specifying the constructive operation to be performed on ide_name's value.

returns

The updated sigma store, where the identifier ide_name is now associated with operation's result.

raises Unbound_variable

with a descriptive message about the variable/identifier, in case the specified ide_name wasn't found inside s.

val cadd : s:sigma -> ide_name:Par_interp.Ast.ide -> value:int -> sigma

Constructive sum (+=): adds value to the value associated to ide_name, and returns the updated sigma store where

ide_name = (ide_name's previous value) + value.

This is a mere wrapper for cassign specifying what the operation parameter function does.

val csub : s:sigma -> ide_name:Par_interp.Ast.ide -> value:int -> sigma

Constructive subtraction (-=): subtracts value from the value associated to ide_name, and returns the updated sigma store where

ide_name = (ide_name's previous value) - value.

This is a mere wrapper for cassign specifying what the operation parameter function does.

val get_value : s:sigma -> ide_name:Par_interp.Ast.ide -> int

Returns the integer value associated to the identifier passed as parameter.

parameter s

The sigma store where the variable containing the value to retrieve is stored.

parameter ide_name

The variable name associated to the value to retrieve in the sigma store.

returns

The integer value associated to ide_name inside s.

raises Unbound_variable

with a descriptive message about the variable/identifier, in case the specified ide_name wasn't found inside s.