Module Serial_interp__Delta

Implementation of the auxiliary store and the functions acting on it.

exception Delta_stack_error of string

val exc_ide_stack_not_found : string -> exn

val exc_if_stack_empty : exn

val exc_while_stack_empty : exn

type ide_stacks =

| Ide_stacks of (Serial_interp.Ast.ide * int list) list

ide_stacks is a type representing a list of tuples (var_name, prev_values_stack); its goal is to keep track of variables' previous values each time a destructive assignment is performed, by pushing the older value onto the respective variable's stack before performing the assignment. By doing so, previous values of any given variable can be retrieved during reverse execution by simply "popping" the value in the head of the list and assigning it to that variable in the sigma store.

val private_push_ide : ide_stk:ide_stacks -> ide_name:Serial_interp.Ast.ide -> value:int -> ide_stacks

Pushes a value onto the specified ide_name variable's stack.

If a tuple (ide_name, ide_stack_list) already exists in ide_stk, value is simply prepended to that list (pushed onto that ide_stack_list): (ide_name, value :: ide_stack_list).

If a tuple for ide_name doesn't exist, a tuple (ide_name, value :: ) gets created and appended to ide_stk.

parameter ide_stk

The list of tuples (variable_name, variable_name_stack_list) representing variables' auxiliary store.

parameter ide_name

The name of the variable which underwent a destructive assignment, for which we want to preserve its previous value by pushing it onto its associated stack.

parameter value

The previous integer value to push onto the stack associated to ide_name.

returns

The updated ide_stk, where value has been pushed onto ide_name's associated stack.

val private_top_ide : ide_stk:ide_stacks -> ide_name:Serial_interp.Ast.ide -> int

Retrieves the integer value from the top of the specified ide_name variable's stack.

parameter ide_stk

The list of tuples (variable_name, variable_name_stack_list).

parameter ide_name

The name of the variable associated to the stack from which we want to retrieve the integer value on its top.

returns

The integer on the top of the stack associated to ide_name.

raises Delta_stack_error

with a descriptive message about the variable/identifier, in case the tuple (ide_name, ide_name_stack_list) wasn't found inside ide_stk.

val private_pop_ide : ide_stk:ide_stacks -> ide_name:Serial_interp.Ast.ide -> ide_stacks

Pops (removes) the integer value from the top of the specified ide_name variable's stack.

parameter ide_stk

The list of tuples (variable_name, variable_name_stack_list).

parameter ide_name

The name of the variable associated to the stack from which we want to retrieve the integer value on its top.

returns

The updated ide_stk, where the element previously located on top of ide_name's associated stack has been removed.

If ide_name's associated stack contained a single element, the whole (ide_name, ide_name_stack_list) pair gets removed from ide_stk.

raises Delta_stack_error

with a descriptive message about the variable/identifier, in case the tuple (ide_name, ide_name_stack_list) wasn't found inside ide_stk.

type if_stack =

| If_stack of bool list

if_stack is a type representing a stack of booleans; its goal is to keep track of which branches have been executed inside Ifthenelse statements, in order to identify the branch to execute during reverse execution by popping values from this stack (true = "then" branch, false = "else" branch.)

val private_push_if : if_stk:if_stack -> value:bool -> if_stack

Pushes value on the top of if_stk stack.

val private_top_if : if_stack -> bool

Retrieves the boolean value from the top of if_stk stack.

raises Delta_stack_error

with a message telling that if_stack is empty, in case the function is applied to an empty if_stk stack.

val private_pop_if : if_stack -> if_stack

Pops (removes) the boolean value from the top of the specified if_stk stack.

raises Delta_stack_error

with a message telling that if_stack is empty, in case the function is applied to an empty if_stk stack.

type while_stack =

| While_stack of bool list

while_stack is a type representing a stack of booleans; its goal is to keep track of how many times the program constituting a While's body has been executed, in order to iterate that program for the same amount of times during reverse execution.

The way it works is that during forward execution false is pushed in while_stack when a While statement is evaluated, and true is pushed at the end of each While's program body whenever a While_end statement is evaluated; by doing so, during reverse execution we keep reiterating While's body as long as true is popped from while_stack whenever While or While_start is evaluated, and execution continues to the statement/instruction preceding While as soon as false is popped: this guarantees that the same amount of iterations of While's body that were performed during forward execution is performed during reverse execution as well.

val private_push_while : while_stk:while_stack -> value:bool -> while_stack

Pushes value on the top of while_stk stack.

val private_top_while : while_stk:while_stack -> bool

Retrieves the boolean value from the top of while_stk stack.

raises Delta_stack_error

with a message telling that while_stack is empty, in case the function is applied to an empty while_stk stack.

val private_pop_while : while_stk:while_stack -> while_stack

Pops (removes) the boolean value from the top of the specified while_stk stack.

raises Delta_stack_error

with a message telling that while_stack is empty, in case the function is applied to an empty while_stk stack.

type delta =

| Delta of ide_stacks * if_stack * while_stack

delta is a type representing the auxiliary store, consisting of:

A list of stacks ide_stacks, where each stack is contained inside a pair (<identifier>, <integer list>) that keeps track of each variable's value: every time a destructive assignment is performed on a certain <ide> variable, its previous value is pushed onto the stack associated to that <ide> variable, in order to preserve it;

A stack if_stack, which keeps track of all the boolean values used as tests inside conditional statements (that is, Ifthenelse statements);

A stack while_stack, whick keeps track of all the boolean values used as tests inside While loops.

val push_ide : d:delta -> ide_name:Serial_interp.Ast.ide -> value:int -> delta

This is a wrapper which applies private_push_ide to the parameter ide_stks inside the specified d delta parameter.

val top_ide : d:delta -> ide_name:Serial_interp.Ast.ide -> int

This is a wrapper which applies private_top_ide to the parameter ide_stks inside the specified d delta parameter.

val pop_ide : d:delta -> ide_name:Serial_interp.Ast.ide -> delta

This is a wrapper which applies private_pop_ide to the parameter ide_stks inside the specified d delta parameter.

val push_if : d:delta -> value:bool -> delta

This is a wrapper which applies private_pop_ide to the parameter if_stk inside the specified d delta parameter.

val top_if : d:delta -> bool

This is a wrapper which applies private_top_if to the parameter if_stk inside the specified d delta parameter.

val pop_if : d:delta -> delta

This is a wrapper which applies private_pop_if to the parameter if_stk inside the specified d delta parameter.

val push_while : d:delta -> value:bool -> delta

This is a wrapper which applies private_push_while to the parameter while_stk inside the specified d delta parameter.

val top_while : d:delta -> bool

This is a wrapper which applies private_top_while to the parameter while_stk inside the specified d delta parameter.

val pop_while : d:delta -> delta

This is a wrapper which applies private_pop_while to the parameter while_stk inside the specified d delta parameter.

val empty_delta : delta

Returns an empty auxiliary store,