Module Serial_interp__Interpreter

Augmented interpreter for the serial version of the while language.

exception Assignment_to_non_var of string

exception Illegal_statement_fwd_execution of string

exception Illegal_statement_rev_execution of string

val sem_int : expr:Serial_interp.Ast.int_expr -> s:Serial_interp.Sigma.sigma -> int

Integer semantics' definition.

parameter expr

The integer expression to evaluate.

parameter s

The sigma store (used in case the expression contains variables whose value must be retrieved.)

returns

The integer value resulting from the expression's evaluation.

val sem_bool : expr:Serial_interp.Ast.bool_expr -> expr:Serial_interp.Sigma.sigma -> bool

Boolean semantics' definition.

parameter expr

The boolean expression to evaluate.

parameter s

The sigma store (used in case the expression contains variables whose value must be retrieved.)

returns

The boolean value resulting from the expression's evaluation.

val assign_var_fwd : e1:Serial_interp.Ast.int_expr -> e2:Serial_interp.Ast.int_expr -> state:Serial_interp.State.state -> Serial_interp.State.state

Performs forward execution's destructive assignment:

Evaluates e1, making sure it's a variable;

Evaluates e2;

Assigns e2's evaluation to the variable specified by e1, and returns the state with the above changes applied (see State.dassign_fwd for details such as pushing the previous value in the delta store before performing the assignment.)

parameter e1

The expression containing the variable to which we must assign the new value.

parameter e2

The expression containing the value to assign.

parameter state

The state where the assignment must be performed.

returns

The new state where the assignment has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val assign_var_rev : e1:Serial_interp.Ast.int_expr -> state:Serial_interp.State.state -> Serial_interp.State.state

Performs reverse execution's destructive assignment:

Evaluates e1, making sure it's a variable ide;

Assigns ide's previous value to ide and returns the state with the above changes applied (see State.dassign_rev for details such as popping ide's previous value from the delta store.)

parameter e1

The expression containing the variable to which we must assign the previous value.

parameter state

The state where the assignment must be performed.

returns

The new state where the assignment has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val cadd : e1:Serial_interp.Ast.int_expr -> e2:Serial_interp.Ast.int_expr -> state:Serial_interp.State.state -> Serial_interp.State.state

Performs constructive sum:

Evaluates e1, making sure it's a variable;

Evaluates e2;

Sums e2's evaluation to the variable specified by e1, and returns the state with the above changes applied (see State.cadd for details.)

parameter e1

The expression containing the variable to which we must add e2's evaluation

parameter e2

The expression containing the value to add to the variable specified by e1.

parameter state

The state where the sum must be performed.

returns

The new state where the sum has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val csub : e1:Serial_interp.Ast.int_expr -> e2:Serial_interp.Ast.int_expr -> state:Serial_interp.State.state -> Serial_interp.State.state

Performs constructive subtraction:

Evaluates e1, making sure it's a variable;

Evaluates e2;

Subtracts e2's evaluation from the variable specified by e1, and returns the state with the above changes applied (see State.csub for details.)

parameter e1

The expression containing the variable from which we must subtract e2's evaluation

parameter e2

The expression containing the value to subtract from the variable specified by e1.

parameter state

The state where the subtraction must be performed.

returns

The new state where the subtraction has been performed.

raises Assignment_to_non_var

if e1's evaluation isn't a variable/identifier.

val if_eval_fwd : b:Serial_interp.Ast.bool_expr -> prg1:Serial_interp.Ast_aug.program_aug -> prg2:Serial_interp.Ast_aug.program_aug -> state:Serial_interp.State.state -> Serial_interp.State.state

Evaluates Ifthenelse statements during forward execution.

The following operations are performed in the given order:

b (Ifthenelse's boolean expression) is evaluated;

prg (The current program where the Ifthenelse statement was encountered) is retrieved from state;

If the result of b's evaluation is true, prg1 (the program associated with the "then" branch) is assigned to if_prg_block for further processing; otherwise (b's evaluation is false), prg2 (the program associated with the "else" branch) is assigned to if_prg_block for further processing;

Ifthenelse's branch to be executed (if_prg_block) gets augmented by calling Augment.aug_if_block and the augmented program is assigned to aug_if_prg_block;

A new state containing aug_if_prg_block as program is returned.

parameter b

Ifthenelse's boolean expression, whose evaluation determines whether to execute prg1 or prg2.

parameter prg1

The program associated with Ifthenelse's "then" branch.

parameter prg2

The program associated with Ifthenelse's "else" branch.

parameter state

The state where the Ifthenelse statement under analysis was encountered.

returns

A new state having as program the augmented Ifthenelse branch to execute.

val if_eval_rev : prg1:Serial_interp.Ast_aug.program_aug -> prg2:Serial_interp.Ast_aug.program_aug -> state:Serial_interp.State.state -> Serial_interp.State.state

Evaluates Ifthenelse statements during reverse execution.

The following operations are performed in the given order:

b_eval (Ifthenelse's boolean evaluation, which was evaluated during forward execution) is retrieved and removed(popped) from Delta.if_stack in the auxiliary delta store;

new_state is created by putting the new delta store with b_eval removed from the top of Delta.if_stack, and moving the program's current statement one statement backward (keep in mind that sem_stmt_rev evaluates the previous statement rather than the current one, therefore if we want Ifthenelse to become the current statement we must move one step back);

prg (The current program where the Ifthenelse statement under evaluation is now the current statement) is retrieved from new_state;

If the result of b_eval is true, prg1 (the program associated with the "then" branch) is assigned to if_prg_block for further processing; otherwise (b's evaluation is false), prg2 (the program associated with the "else" branch) is assigned to if_prg_block for further processing;

Ifthenelse's branch to be executed (if_prg_block) gets augmented by calling Augment.aug_if_block; the current statement in the augmented program is set to the last one by feeding it to Program.last_stmt (we are performing reverse execution, so we must begin from the program's last statement) and the result is assigned to aug_if_prg_block;

A new state containing aug_if_prg_block as program is returned.

parameter prg1

The program associated with Ifthenelse's "then" branch.

parameter prg2

The program associated with Ifthenelse's "else" branch.

parameter state

The state where the Ifthenelse statement under analysis was encountered.

returns

A new state having as program the augmented Ifthenelse branch to execute (starting from the program's last statement), as well as an updated Delta.if_stack in the sigma store with the top value removed from the stack.

val while_eval_fwd : b_expr:Serial_interp.Ast.bool_expr -> while_prg:Serial_interp.Ast_aug.program_aug -> state:Serial_interp.State.state -> Serial_interp.State.state

Evaluates While statements during forward execution.

The following operations are performed in the given order:

b_expr (While's boolean expression) is evaluated and assigned to b_eval;

new_state is created by pushing false onto Delta.while_stack; this step is necessary since during forward execution we have the following two scenarios:

While's first b_expr evaluation was false (While's program block was skipped): since no true values were pushed onto Delta.while_stack, the false value we pushed gets popped during reverse execution and While's program block gets skipped again;

While's b_expr evaluates to true for a given n amount of times; n true values get pushed onto Delta.while_stack, therefore during reverse execution we pop those true values which make the interpreter iterate While's program block n times, at which point the false value we pushed gets popped and While's reverse iteration stops.

The program having the statement following the current While as the current statement is assigned to prg_while_next_stmt;

if b_eval is true, it means we must execute the program block inside While at least once, therefore while_prg gets augmented by calling Augment.aug_while_block and assigned to prg_to_execute; otherwise (b_eval = false), While's program block gets skipped and prg_while_next_stmt is assigned to prg_to_execute;

A state having prg_to_execute as program and new_state as state is returned.

parameter b_expr

While's boolean expression.

parameter while_prg

The program associated to While's body;

parameter state

The state where the While statement has been encountered.

returns

A state having:

as a program: the augmented while_prg if b_expr's evaluation is true, or the program having the statement following the While under evaluation as current statement if b_expr's evaluation is false;

as sigma store: the same as the state parameter;

as auxiliary delta store: state's delta store, with false pushed onto Delta.while_stack.

val while_eval_rev : b_expr:Serial_interp.Ast.bool_expr -> while_prg:Serial_interp.Ast_aug.program_aug -> state:Serial_interp.State.state -> Serial_interp.State.state

Evaluates While statements during reverse execution.

The following operations are performed in the given order:

The boolean value on the top of Delta.while_stack gets retrieved and assigned to b_eval;

new_state is created from state by removing (popping) Delta.while_stack's top value and setting as the current statement the While statement (keep in mind that sem_stmt_rev evaluates the previous statement rather than the current one, therefore if we want While to become the current statement we must move one step back);

The program having the statement preceding the current While as the current statement is assigned to prg_while_prev_stmt;

if b_eval is true, it means we must execute the program block inside While at least once, therefore while_prg gets augmented by calling Augment.aug_while_block and assigned to prg_to_execute; otherwise (b_eval = false), While's program block gets skipped and prg_while_next_stmt is assigned to prg_to_execute;

A state having prg_to_execute as program and new_state as state is returned.

parameter b_expr

While's boolean expression.

parameter while_prg

The program associated to While's body;

parameter state

The state where the While statement has been encountered.

returns

A state having:

as a program: the augmented while_prg if b_expr's evaluation is true, or the program having the statement following the While under evaluation as current statement if b_expr's evaluation is false;

as sigma store: the same as the state parameter;

as auxiliary delta store: state's delta store, with false pushed onto Delta.while_stack.

val while_start_eval : while_stmt_in_outer_prg:Serial_interp.Ast_aug.program_aug -> state:Serial_interp.State.state -> Serial_interp.State.state

Evaluates While_start statements during reverse execution.

NOTE: While_start statements can be evaluated ONLY during reverse execution, since they mark the beginning of a While's program body; if they get evaluated during forward execution, then there's a problem (exception Illegal_statement_fwd_execution would be raised), but it should never happen anyway.

The following operations are performed in the given order:

The boolean value on the top of Delta.while_stack gets retrieved and assigned to b_eval;

new_state is created from state by removing (popping) Delta.while_stack's top value (which we just assigned to b_eval in the previous step);

if b_eval is true, it means we must perform another reverse iteration of the While program block we're currently executing starting from the last statement,therefore we set the current statement to the last statement of the current program block (by calling Program.last_stmt) and the result is assigned to prg_to_execute. Otherwise (b_eval is false), it means the last iteration of While's program has been performed and we need to jump out of the loop, therefore while_stmt_in_outer_prg (which is While_start's parameter consisting in the outer program having as current statement the While statement containing the program block which is currently being executed) gets assigned to prg_to_execute;

A state having prg_to_execute as program and new_state as state is returned.

parameter while_stmt_in_outer_prg

The program having the While statement whose body is being executed, where execution will continue in case reverse iteration must be stopped due to a false value popped from Delta.while_stack.

parameter state

The current state, where the While_start statement has been encountered.

returns

A state having:

as a program: the current program having the current statement set to the program's last statement if the value retrieved from Delta.while_stack is true, or while_stmt_in_outer_prg if the value retrieved from Delta.while_stack is false;

as sigma store: the same as the state parameter;

as auxiliary delta store: state's delta store, with the retrieved boolean value b_eval popped(removed) from the top of Delta.while_stack.

val while_end_eval : b_expr:Serial_interp.Ast.bool_expr -> while_next_stmt_in_outer_prg:Serial_interp.Ast_aug.program_aug -> state:Serial_interp.State.state -> Serial_interp.State.state

Evaluates While_end statements during forward execution.

NOTE: While_end statements can be evaluated ONLY during forward execution, since they mark the end of a While's program body; if they get evaluated during reverse execution, then there's a problem (exception Illegal_statement_rev_execution would be raised), but it should never happen anyway.

The following operations are performed in the given order:

b_expr (While's boolean expression) is evaluated and assigned to b_eval;

new_state is created from state by pushing true onto Delta.while_stack (We're at the end of a While program iteration which has has been executed, therefore we must push true to mark the fact an iteration has been executed when we reverse execute the program, regardless of b_eval's value);

if b_eval is true, it means we must perform another forward iteration of the While program block we're currently executing starting from the first statement,therefore we set the current statement to the first statement of the current program block (by calling Program.while_block_1st_stmt) and the result is assigned to prg_to_execute. Otherwise (b_eval is false), it means the last iteration of While's program has been performed and we need to jump out of the loop, therefore while_next_stmt_in_outer_prg (which is While_end's parameter consisting in the outer program having as current statement the statement following the While statement containing the program block which is currently being executed) is assigned to prg_to_execute;

A state having prg_to_execute as program and new_state as state is returned.

parameter b_expr

While_end's boolean expression, inherited from the containing While statement.

parameter while_next_stmt_in_outer_prg

The program having the While statement whose body is being executed, where execution will continue to While's next statement in case iteration must be stopped due to b_expr evaluating to false.

parameter state

The current state, where the While_start statement has been encountered.

returns

A state having:

as a program: the current program having the current statement set to the program's first statement if b_expr's evaluation is true, or while_next_stmt_in_outer_prg if b_expr's evaluation is false;

as sigma store: the same as the state parameter;

as auxiliary delta store: state's delta store, with true pushed onto Delta.while_stack.

val sem_stmt_fwd : Serial_interp.State.state -> Serial_interp.State.state

Given a state, performs a single evaluation step in forward execution mode and returns the resulting state.

val sem_stmt_rev : Serial_interp.State.state -> Serial_interp.State.state

Given a state, performs a single evaluation step in reverse execution mode and returns the resulting state.

val sem_prg_fwd : Serial_interp.State.state -> Serial_interp.State.state

Given a state, evaluates all statements until the end of the program in forward execution mode and returns the resulting state.

val sem_prg_rev : Serial_interp.State.state -> Serial_interp.State.state

Given a state, evaluates all statements until the beginning of the program in reverse execution mode and returns the resulting state.

val sem_prg_fwd_steps : Serial_interp.State.state -> int -> Serial_interp.State.state

Given a state curr_state and a specified num_steps integer, let remaining_stmts be the number of statements between the current statement in curr_state's program and the Program_end boundary statement; then the function performs min(remaining_stmts, num_steps) statement evaluations in forward execution mode and returns the resulting state.

If num_steps <= 0, curr_state is returned unaltered.

val sem_prg_rev_steps : Serial_interp.State.state -> int -> Serial_interp.State.state

Given a state curr_state and a specified num_steps integer, let remaining_stmts be the number of statements between the current statement in curr_state's program and the Program_start boundary statement; then the function performs min(remaining_stmts, num_steps) statement evaluations in reverse execution mode and returns the resulting state.

If num_steps <= 0, curr_state is returned unaltered.

val sem_prg_steps : Serial_interp.State.state -> int -> Serial_interp.State.state

Given a state curr_state and a specified num_steps integer, it acts as an interface allowing both forward and reverse execution by calling

sem_prg_fwd_steps (forward execution), if num_steps >= 0;

sem_prg_rev_steps (reverse execution), if num_steps < 0 (in this case, num_steps's sign gets inverted to obtain the positive amount of steps to execute to pass to sem_prg_rev_steps as parameter.)