Module Serial_interp__Program

Implementation of the functions acting on Ast_aug.program_aug.

exception No_previous_statements

exception No_next_statements

exception If_start_not_found

exception While_start_not_found

val prepend_stmt : stmt_aug:Serial_interp.Ast_aug.stmt_aug -> Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

Given a program, adds stmt_aug as the program's first statement/instruction.

val append_stmt : Serial_interp.Ast_aug.stmt_aug -> Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

Given a program, adds stmt_aug as the program's last statement/instruction.

val prev_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

Given a program prg, returns:

A new program where prg's previous statement becomes the current statement, or

prg, if prg's current statement is the first one (there are no previous statements.)

val next_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

Given a program prg, returns:

A new program where prg's next statement becomes the current statement, or

prg, if prg's current statement is the last one (there are no next statements.)

val last_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

Given a program prg, returns a new program where prg's last statement becomes the current statement.

val if_block_1st_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

"Rewinds" a program associated to an Ifthenelse's "then" or "else" branch back to the first statement (that is, the statement coming after If_start.)

raises If_start_not_found

if If_start wasn't found as the program's first statement (e.g. the function is applied to an Ifthenelse program block that wasn't augmented, or wasn't applied to an Ifthenelse block at all.)

val while_block_1st_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.program_aug

"Rewinds" a program associated to a While's block back to the first statement (that is, the statement coming after While_start.)

raises While_start_not_found

if While_start wasn't found as the program's first statement (e.g. the function is applied to a While program block that wasn't augmented, or wasn't applied to a While block at all.)

val get_prev_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.stmt_aug

Returns the statement preceding the current statement inside the program.

raises No_previous_statements

if the previous statements' list is empty (e.g. the given program's current statement is the first one.)

val get_curr_stmt : Serial_interp.Ast_aug.program_aug -> Serial_interp.Ast_aug.stmt_aug

Returns the current statement inside the program.