Module Serial_interp__State

Implementation of the state and the functions acting on it.

type state =

| State of Serial_interp.Ast_aug.program_aug * Serial_interp.Sigma.sigma * Serial_interp.Delta.delta

state is a type consisting of a tuple including:

program_aug (the list of statements constituting the program);

The program store (sigma);

The auxiliary store (delta).

val dassign_fwd : ide_name:Serial_interp.Ast.ide -> new_value:int -> state -> state

Executes a destructive assignment in forward execution mode, assigning new_value to the variable ide_name.

In order to preserve ide_name's current value, curr_value is retrieved from the sigma store and pushed onto ide_name's stack in the auxiliary delta store; at that point, ide_name's value in the sigma store can be overwritten with new_value, and the resulting new state is returned.

val dassign_rev : ide_name:Serial_interp.Ast.ide -> state -> state

Executes a destructive assignment in reverse execution mode.

ide_name's previous value (retrieved_value) is retrieved from ide_name's stack in the auxiliary delta store, and retrieved_value gets popped (removed) from ide_name's stack; at that point, retrieved_value is assigned to ide_name in the sigma store, and the resulting new state is returned.

val cadd : Serial_interp.Ast.ide -> int -> state -> state

This is a wrapper which applies Sigma.cadd to the sigma store contained inside the specified state parameter, and returns the resulting state.

val csub : Serial_interp.Ast.ide -> int -> state -> state

This is a wrapper which applies Sigma.csub to the sigma store contained inside the specified state parameter, and returns the resulting state.

val push_if : bool -> state -> state

This is a wrapper which applies Delta.push_if to the auxiliary delta store contained inside the specified state parameter, and returns the resulting state.

val top_if : state -> bool

This is a wrapper which applies Delta.top_if to the auxiliary delta store contained inside the specified state parameter, and returns the resulting state.

val pop_if : state -> state

This is a wrapper which applies Delta.pop_if to the auxiliary delta store contained inside the specified state parameter, and returns the resulting state.

val push_while : bool -> state -> state

This is a wrapper which applies Delta.push_while to the auxiliary delta store contained inside the specified state parameter, and returns the resulting state.

val top_while : state -> bool

This is a wrapper which applies Delta.top_while to the auxiliary delta store contained inside the specified state parameter, and returns the resulting state.

val pop_while : state -> state

This is a wrapper which applies Delta.pop_while to the auxiliary delta store contained inside the specified state parameter, and returns the resulting state.

val get_program : state -> Serial_interp.Ast_aug.program_aug

Returns the program contained inside the specified state parameter.

val set_program : Serial_interp.Ast_aug.program_aug -> state -> state

Given a program p and a state state as parameters, returns a state consisting of the given state with p replacing the previously contained program.

val get_sigma : state -> Serial_interp.Sigma.sigma

Returns the sigma store contained inside the specified state parameter.

val init : Serial_interp.Ast_aug.program_aug -> Serial_interp.Sigma.sigma -> state

Takes a program and an initial sigma store as parameters, and returns a state with

The given program;

The given sigma store;

An empty auxiliary (delta) store.

val prepend_stmt : Serial_interp.Ast_aug.stmt_aug -> state -> state

This is a wrapper which applies Program.prepend_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val append_stmt : Serial_interp.Ast_aug.stmt_aug -> state -> state

This is a wrapper which applies Program.append_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val prev_stmt : state -> state

This is a wrapper which applies Program.prev_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val get_prev_stmt : state -> Serial_interp.Ast_aug.stmt_aug

This is a wrapper which applies Program.get_prev_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val get_curr_stmt : state -> Serial_interp.Ast_aug.stmt_aug

This is a wrapper which applies Program.get_curr_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val next_stmt : state -> state

This is a wrapper which applies Program.next_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val last_stmt : state -> state

This is a wrapper which applies Program.last_stmt to the program contained inside the specified state parameter, and returns the resulting state.

val is_prg_at_start : state -> bool

Returns a boolean value indicating whether the program's current statement/instruction in the given state is the main program's first statement/instruction.

val is_prg_at_end : state -> bool

Returns a boolean value indicating whether the program's current statement/instruction in the given state is the main program's last statement/instruction.